requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.11
gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.2
openai>=1.0.0
stripe>=7.0.0
resend>=2.0.0
Flask-Limiter>=3.0.0
markdown>=3.5.0
pyyaml>=6.0
